Updates structure
- Renames index -> Helmet
- Adds index file for exports
